Fold offset and containment decls into boolean2.h; drop tiny headers
