+++ src/cross_section/boolean2/boolean2.h
@@ -35,5 +35,15 @@ Polygons Simplify(const Polygons& in, double eps, double tolerance = 0.0);
 Polygons Boolean2D(const Polygons& a, const Polygons& b, OpType op,
                    double eps = 0.0, double tolerance = 0.0);
 
+// Polygon offset backing CrossSection::Offset.
+enum class OffsetJoinType { Square, Round, Miter, Bevel };
+Polygons Offset(const Polygons& in, double delta, OffsetJoinType jt,
+                double miterLimit = 2.0, int circularSegments = 0,
+                double tolerance = 0.0);
+
+// Group regularized simple loops into outer-ring components with their
+// directly contained holes, backing CrossSection::Decompose.
+std::vector<Polygons> DecomposeByContainment(const Polygons& polys);
+
 }  // namespace boolean2
 }  // namespace manifold
