+++ src/cross_section/boolean2/offset.cpp
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "offset.h"
-
 #include <algorithm>
 #include <cmath>
 #include <vector>
